JAMES-1902 Extract 'mailrepository' cassandra implementation into its own maven module
