+++ server/container/guice/jpa-common-guice/src/main/java/org/apache/james/modules/data/JPAMailRepositoryModule.java
@@ -25,17 +25,12 @@ import org.apache.james.mailrepository.api.Protocol;
 import org.apache.james.mailrepository.file.FileMailRepository;
 import org.apache.james.mailrepository.jpa.JPAMailRepositoryUrlStore;
 import org.apache.james.mailrepository.memory.MailRepositoryStoreConfiguration;
-import org.apache.james.modules.server.MailStoreRepositoryModule;
 
 import com.google.common.collect.ImmutableList;
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
 
 public class JPAMailRepositoryModule extends AbstractModule {
-    private static final MailRepositoryStoreConfiguration.Item FILE_MAILREPOSITORY_DEFAULT_DECLARATION = new MailRepositoryStoreConfiguration.Item(
-        ImmutableList.of(new Protocol("file")),
-        FileMailRepository.class.getName(),
-        new BaseHierarchicalConfiguration());
 
     @Override
     protected void configure() {
@@ -43,6 +38,10 @@ public class JPAMailRepositoryModule extends AbstractModule {
 
         bind(MailRepositoryUrlStore.class).to(JPAMailRepositoryUrlStore.class);
 
-        bind(MailStoreRepositoryModule.DefaultItemSupplier.class).toInstance(() -> FILE_MAILREPOSITORY_DEFAULT_DECLARATION);
+        bind(MailRepositoryStoreConfiguration.Item.class)
+            .toProvider(() -> new MailRepositoryStoreConfiguration.Item(
+                ImmutableList.of(new Protocol("file")),
+                FileMailRepository.class.getName(),
+                new BaseHierarchicalConfiguration()));
     }
 }
