+++ server/container/guice/pom.xml
@@ -56,6 +56,7 @@
         <module>mailbox</module>
         <module>mailbox-plugin-deleted-messages-vault-guice</module>
         <module>mailbox-plugin-spamassassin</module>
+        <module>mailrepository-cassandra</module>
         <module>mailet</module>
         <module>memory-guice</module>
         <module>onami</module>
@@ -239,6 +240,11 @@
             <dependency>
                 <groupId>${james.groupId}</groupId>
                 <artifactId>james-server-jpa-smtp-common-guice</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${james.groupId}</groupId>
+                <artifactId>mailrepository-cassandra</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
